Changed quiesce event to check for si_usecount
This is a better method than atomic as is uses the dev_ref/rel that I
previously implemented.

I had to make cdev public in the Cdev<T> struct in order to access and deref it in the
quiesce method, this propabily isnt ideal but it gets the job done.